backend/pyproject.toml
[project]
name = "medsearch-backend"
version = "1.0.0"
description = "MedSearch AI Backend - Multi-agent medical research assistant"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "python-multipart==0.0.9",
    "langgraph==0.2.28",
    "langchain==0.3.1",
    "langchain-core==0.3.6",
    "langchain-community==0.3.1",
    "langchain-google-vertexai==2.0.4",
    "elasticsearch==8.15.1",
    "google-cloud-aiplatform==1.68.0",
    "google-auth==2.34.0",
    "redis==5.0.8",
    "aiosqlite==0.20.0",
    "aiohttp==3.10.5",
    "httpx==0.27.2",
    "pydantic==2.9.1",
    "pydantic-settings==2.5.2",
    "slowapi==0.1.9",
    "python-json-logger==2.0.7",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
    "pytest-mock==3.14.0",
    "ruff==0.6.8",
    "mypy==1.11.2",
    "black==24.8.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ['py311']

